src/test/java/com/attornatus/controllerTest/PessoaControllerTest.java
package com.attornatus.controllerTest;

import com.attornatus.controller.PessoaController;
import com.attornatus.entities.Pessoa;
import com.attornatus.service.EnderecoService;
import com.attornatus.service.PessoaService;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
import org.junit.jupiter.api.Test;
import org.junit.runner.RunWith;
import org.mockito.Mockito;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;

import java.time.LocalDate;

@RunWith(SpringRunner.class)
@WebMvcTest(PessoaController.class)
public class PessoaControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private PessoaService pessoaService;

    @MockBean
    private EnderecoService enderecoService;
    @Test
    public void testCriarPessoa() throws Exception {
        Pessoa pessoa = new Pessoa();
        pessoa.setNome("Alice");
        pessoa.setDataNascimento(LocalDate.of(1990, 5, 15));

        Mockito.when(pessoaService.criarPessoa(Mockito.any(Pessoa.class))).thenReturn(pessoa);

        mockMvc.perform(MockMvcRequestBuilders.post("/api/pessoas")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content(asJsonString(pessoa)))
                .andExpect(MockMvcResultMatchers.status().isCreated())
                .andExpect(MockMvcResultMatchers.jsonPath("$.nome").value("Alice"));
    }

    @Test
    public void testConsultarPessoaPorId() throws Exception {
        Long pessoaId = 1L;
        Pessoa pessoa = new Pessoa();
        pessoa.setId(pessoaId);
        pessoa.setNome("Bob");
        pessoa.setDataNascimento(LocalDate.of(1985, 3, 20));

        Mockito.when(pessoaService.consultarPessoaPorId(pessoaId)).thenReturn(pessoa);

        mockMvc.perform(MockMvcRequestBuilders.get("/api/pessoas/{pessoaId}", pessoaId)
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(MockMvcResultMatchers.status().isOk())
                .andExpect(MockMvcResultMatchers.jsonPath("$.nome").value("Bob"));
    }

    private static String asJsonString(final Object obj) {
        try {
            final ObjectMapper mapper = new ObjectMapper();
            mapper.registerModule(new JavaTimeModule());
            return mapper.writeValueAsString(obj);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }
}

